pyproject.toml
[project]
name = "khrspeechcommander"
version = "0.1.0"
description = "Controller of KHR-3HV by speech"
readme = "README.md"
requires-python = "==3.12.8"
dependencies = [
    "flet[all]==0.26.0",
    "mecab-python3>=1.0.10",
    "numpy>=2.2.3",
    "pyaudio>=0.2.14",
    "speechrecognition>=3.14.1",
]
